optimized adjacent blocks getters
After profiling, these methods were the bottleneck.
This was mostly because they used HashSets which were not necessary but added a lot of overhead.
Also, they iterated over the returned collection to remove null elements in all cases, while this was actually only necessary when wrapping is set to CUT, which does not happen so often.